estoy en unas modificaciones en la estructura del dashboard en subcomponentes
