+++ frontend/src/components/Dashboard/subcomponents/Totals.tsx
@@ -1,7 +1,5 @@
+
 import { Box, Card, CardContent, Typography } from '@mui/material';
-import { PieChartComponent } from './PieChart';
-import expenses from '../../mocks/expenses.mock.json';
-import DashedLineChart from './DashedLineChart';
 import { theme } from "@/constants/theme";
 
 // Función que devuelve el color según el valor numérico
@@ -11,97 +9,12 @@ const getColorByValue = (value: number): string => {
   return 'text.primary'; // negro (color por defecto)
 };
 
-export const Dashboard = () => {
-  // Valores ejemplo, podés traerlos de donde quieras
-  const gastosHoy = 0;
-  const gastoSemana = -100;
-  const gastoMes = -100;
-  const balance = 1000;
+export const Totals = () => {
+    const totalGastos = -100;
+    const totalIngresos = 300;
+    const balanceResumen = 200;
 
-  const totalGastos = -100;
-  const totalIngresos = 300;
-  const balanceResumen = 200;
-
-  return (
-    <>
-      <Box
-        sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}
-      >
-        <Box
-          sx={{
-            display: 'flex',
-            flexWrap: 'wrap',
-            gap: 2,
-            width: '70%',
-            p: 2,
-            border: '1px solid yellowgreen',
-          }}
-        >
-          {/* Gasto Hoy */}
-          <Box sx={{ flex: { xs: '35%', md: '20%' } }}>
-            <Card sx={{ height: '100%', borderRadius:'10px', background: theme.colors.inputBg, color:'white' }}>
-              <CardContent>
-                <Typography variant="subtitle2">Gasto Hoy</Typography>
-                <Typography
-                  variant="h5"
-                  color={getColorByValue(gastosHoy)}
-                  fontWeight="bold"
-                >
-                  ${gastosHoy.toFixed(2)}
-                </Typography>
-              </CardContent>
-            </Card>
-          </Box>
-
-          {/* Gasto Semana */}
-          <Box sx={{ flex: { xs: '35%', md: '20%' } }}>
-            <Card sx={{ height: '100%', borderRadius:'10px', background: theme.colors.inputBg, color:'white' }}>
-              <CardContent>
-                <Typography variant="subtitle2">Gasto Semana</Typography>
-                <Typography
-                  variant="h5"
-                  color={getColorByValue(gastoSemana)}
-                  fontWeight="bold"
-                >
-                  ${Math.abs(gastoSemana).toFixed(2)}
-                </Typography>
-              </CardContent>
-            </Card>
-          </Box>
-
-          {/* Gasto Mes */}
-          <Box sx={{ flex: { xs: '35%', md: '20%' } }}>
-            <Card sx={{ height: '100%', borderRadius:'10px', background: theme.colors.inputBg, color:'white' }}>
-              <CardContent>
-                <Typography variant="subtitle2">Gasto Mes</Typography>
-                <Typography
-                  variant="h5"
-                  color={getColorByValue(gastoMes)}
-                  fontWeight="bold"
-                >
-                  ${Math.abs(gastoMes).toFixed(2)}
-                </Typography>
-              </CardContent>
-            </Card>
-          </Box>
-
-          {/* Balance */}
-          <Box sx={{ flex: { xs: '35%', md: '20%' } }}>
-            <Card sx={{ height: '100%', borderRadius:'10px', background: theme.colors.inputBg, color:'white' }}>
-              <CardContent>
-                <Typography variant="subtitle2">Balance</Typography>
-                <Typography
-                  variant="h5"
-                  color={getColorByValue(balance)}
-                  fontWeight="bold"
-                >
-                  ${balance.toFixed(2)}
-                </Typography>
-              </CardContent>
-            </Card>
-          </Box>
-        </Box>
-
+    return (
         <Box
           sx={{
             display: 'flex',
@@ -217,37 +130,5 @@ export const Dashboard = () => {
             </Card>
           </Box>
         </Box>
-
-        <Box
-          sx={{
-            display: 'flex',
-            gap: 2,
-            width: '70%',
-            p: 2,
-            border: '1px solid violet',
-            mt: 2,
-            flexDirection: { xs: 'column', md: 'row' },
-          }}
-        >
-          {/* Gráfico circular */}
-          <Box sx={{ 
-              width: { xs: '100%', md: '50%' },
-           }}>
-            <Card sx={{ height: '100%', background: theme.colors.inputBg, color:'white'}}>
-             <PieChartComponent expenses={expenses} />
-            </Card>
-          </Box>
-
-          {/* Gráfico de barras */}
-          <Box sx={{ 
-            width: { xs: '100%', md: '50%' },
-          }}>
-           <Card sx={{ height: '100%', background: theme.colors.inputBg, color:'white' }}>
-             <DashedLineChart />
-            </Card>
-          </Box>
-        </Box>
-      </Box>
-    </>
-  );
-};
+    )
+}
